fixed new location for insertion of relations
